+++ src/sorts/MergeSort.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 public class MergeSort extends SortBase {
     @Override
     public void sort() {
-        double[] sortedArray = Arrays.copyOf(this.arr, this.arr.length);
+        int[] sortedArray = Arrays.copyOf(this.arr, this.arr.length);
         sortedArray = this.sortRecursively(sortedArray, (int) Math.pow(2, Math.ceil(Math.log((float) sortedArray.length))));
         this.print(sortedArray);
     }
@@ -13,20 +13,20 @@ public class MergeSort extends SortBase {
     /**
      * 这个地方第一次尝试了从更长的子长度向上递归，结果发现，并不能作为
      */
-    private double[] sortRecursively(double[] sortedArray, int subArrayLength) {
+    private int[] sortRecursively(int[] sortedArray, int subArrayLength) {
         if (subArrayLength == 1) {
-            double[] recursiveArray = Arrays.copyOf(sortedArray, sortedArray.length);
+            int[] recursiveArray = Arrays.copyOf(sortedArray, sortedArray.length);
             for (int i = 1; i < sortedArray.length; i += 2) {
                 if (sortedArray[i] < sortedArray[i - 1]) {
-                    double temp = recursiveArray[i];
+                    int temp = recursiveArray[i];
                     recursiveArray[i] = recursiveArray[i - 1];
                     recursiveArray[i - 1] = temp;
                 }
             }
             return recursiveArray;
         }
-        double[] recursiveArray = new double[sortedArray.length];
-        double[] newArray = this.sortRecursively(sortedArray, (int) Math.pow(2, Math.ceil(Math.log((float) subArrayLength)) - 1));
+        int[] recursiveArray = new int[sortedArray.length];
+        int[] newArray = this.sortRecursively(sortedArray, (int) Math.pow(2, Math.ceil(Math.log((float) subArrayLength)) - 1));
         int startPoint = 0;
         while (startPoint < newArray.length) {
             merge(subArrayLength, recursiveArray, newArray, startPoint);
@@ -35,7 +35,7 @@ public class MergeSort extends SortBase {
         return recursiveArray;
     }
 
-    private void merge(int subArrayLength, double[] recursiveArray, double[] sortedArray, int startPoint) {
+    private void merge(int subArrayLength, int[] recursiveArray, int[] sortedArray, int startPoint) {
         int firstPoint = startPoint;
         int secondPoint = startPoint + subArrayLength;
         int recursivePoint = startPoint;
