+++ src/sorts/SelectionSort.java
@@ -5,10 +5,10 @@ import java.util.Arrays;
 public class SelectionSort extends SortBase {
     @Override
     public void sort() {
-        double[] sortedArray = Arrays.copyOf(this.arr, this.arr.length);
+        int[] sortedArray = Arrays.copyOf(this.arr, this.arr.length);
 
         for (int i = 0; i < sortedArray.length; i++) {
-            double min = sortedArray[i];
+            int min = sortedArray[i];
             int minIndex = i;
             for (int j = i + 1; j < sortedArray.length; j++) {
                 if (sortedArray[j] < min) {
@@ -16,7 +16,7 @@ public class SelectionSort extends SortBase {
                     minIndex = j;
                 }
             }
-            double temp = sortedArray[i];
+            int temp = sortedArray[i];
             sortedArray[i] = sortedArray[minIndex];
             sortedArray[minIndex] = temp;
         }
